refactored code to support uint16_t for arrays

+++ bubble_sort/bubble_sort.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
+#include <stdint.h>
 
 unsigned int ARRAY_LENGTH = 10;
 
 /* checks if a given array is sorted */
-int check_sorted(unsigned int *nums, unsigned int nums_len) {
+int check_sorted(uint16_t *nums, unsigned int nums_len) {
   for (unsigned int i = 0; i < nums_len - 1; i++) {
     if (nums[i] > nums[i + 1]) {
       return 0;
@@ -15,7 +16,7 @@ int check_sorted(unsigned int *nums, unsigned int nums_len) {
 }
 
 /* prints an array in human-readable output to standard output */
-void print_arr(unsigned int *nums, unsigned int nums_len) {
+void print_arr(uint16_t *nums, unsigned int nums_len) {
   printf("[\n");
   for (unsigned int i = 0; i < nums_len; i++) {
     printf("  %d", nums[i]);
@@ -28,14 +29,14 @@ void print_arr(unsigned int *nums, unsigned int nums_len) {
 }
 
 /* swaps two elements in an array */
-void swap(unsigned int *nums, unsigned int idx1, unsigned int idx2) {
-  unsigned int temp = nums[idx1];
+void swap(uint16_t *nums, unsigned int idx1, unsigned int idx2) {
+  uint16_t temp = nums[idx1];
   nums[idx1] = nums[idx2];
   nums[idx2] = temp;
 }
 
 /* preforms bubble-sort in place on an array */
-void bubble_sort(unsigned int *nums, unsigned int nums_len) {
+void bubble_sort(uint16_t *nums, unsigned int nums_len) {
   for (unsigned int i = 0; i < nums_len; i++) {
     for (unsigned int j = 0; j < nums_len - i - 1; j++) {
       if (nums[j] > nums[j + 1]) {
@@ -46,7 +47,7 @@ void bubble_sort(unsigned int *nums, unsigned int nums_len) {
 }
 
 int main() {
-  unsigned int *arr = (unsigned int *)malloc(ARRAY_LENGTH * sizeof(int));
+  uint16_t *arr = (uint16_t *)malloc(ARRAY_LENGTH * sizeof(uint16_t));
   srand(time(NULL));
 
   for (unsigned int i = 0; i < ARRAY_LENGTH; i++) {
